requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cloudinary==1.29.0
dj-database-url==0.5.0
Django==4.0.6
django-cors-headers==3.13.0
django-heroku==0.3.1
django-resized==1.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
Pillow==9.2.0
psycopg2==2.9.3
PyJWT==2.4.0
python-decouple==3.6
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.2.0
